Switch to MurmurHash64A in hash log
Basically, update DiskInterface::HashFile() from MurmurHash2() to
MurmurHash64A().

+++ src/disk_interface.h
@@ -63,7 +63,7 @@ struct DiskInterface: public FileReader {
   virtual int RemoveFile(const string& path) = 0;
 
   /// the hash type returned for HashFile()
-  typedef unsigned int hash_t;
+  typedef uint64_t hash_t;
   /// Hash a file's content
   hash_t HashFile(const string&, string* err);
